Add SCION Rust library
Today we have made our endhost library written in Rust public.

Currently, we have a working UDP socket as well as a raw socket to send and receive SCMP messages (similar features to the snet library). We also have tools to cache and refresh paths via the SCION daemon.

We will continue working on the repository, for example to implement better path management (similar to the pan library), increase resilience, and add a QUIC/SCION socket.